Closes #1737: Datatype misalignment fix
Submitted by bosilca, Reviewed by tdd, RM-approved by bbenton

r20133:
Fix an issue where the datatype description introduce a buffer
misalignment. Because some architectures (read SPARC64) require
aligned accesses, we increase the storage space when we pack a
datatype description to keep the fields aligned. This has to
be done on both sided in order to be consistent.

r20134:
Some 64 bits architectures support pointer aligned on 4 bytes 
(where the sizeof(long) is 8). On such architectures dont 
assert if the datatype representation is not aligned on 64 bits.

r20140:
TYPE shouldn't be surrounded by parens because it causes
issues for some versions of gcc when the construct
{{{ a = ((int)) sizeof(b) }}} comes up...