Closes #1725: Remove usage of the freelist _WAIT function from the BTLs
Submitted by bosilca, Reviewed by various, RM-approved by bbenton

Refs #1713

r20124:
This is a temporary fix for the deadlock problem over MX. The real 
problem seems to come from the free list, but due to lack of time to 
understand it completely, I provide this fix. Basically, there is no 
waiting in the MX BTL anymore, if we cannot allocate a fragment we 
rely on the PML to take the corrective actions.

r20143:
Following ticket #1725, when a free list item can not be allocated, return 
the error to the upper layer and let it deal with the problem

r20149: Update the Elan BTL to fulfill the requirements for #1713.

r20150: Update the sm BTL to fulfill the requirements for #1713.

r20151: Update the tcp BTL to fulfill the requirements for #1713.

r20152: Update the self BTL to fulfill the requirements for #1713.

r20153: Update the template BTL to fulfill the requirements for #1713.

r20231: Update the SCTP BTL regarding ticket #1725.