Closes #1754: configure fix
Submitted by jsquyres, Reviewed by rhc, RM-approved by bbenton

r20045:
If someone specifies a directory for a config option, then we
should respect that designation and error out if the directory
isn't found. Implemented first for the --with-openib options.

r20289:
Add a missing comma, which significantly changes the effect of this macro :-)