Closes #1873: libnuma-and-openib-fix.patch
Submitted by jsquyres, reviewed by bbarrett, RM-approved by bbenton

The attached patch fixes when you specify "--with-libnuma" to configure (noted on the users list: http://www.open-mpi.org/community/lists/users/2009/04/8853.php). It also makes a help message for the --with-openib a bit more clear.

(this is a patch because we already use something different on the trunk for libnuma m4 configuration)