Fixes #1971
r21619

RDMA CM doesn't retry if a packet is dropped, just timesout during
route discovery, which results in a timeout and we don't recover.
Instead, try to recover a couple of times by retrying.

r21620

Drain pipe from service thread to main thread during shutdown.  By
this point, the event engine has been shut down until btl finalization
is done, so opal_progress in the wait loop is not an option - we have
to drain from inside the btl.

Clean up the looping structure for the finalize routine

Update copyrights.