Closes #2024: SLURM nodelists are out of order
Submitted by rhc, reviewed by jsquyres, RM-approved by bbenton

SLURM apparently re-orders the nodelist given to it at launch, thus
causing OMPI to mis-compute the name of the daemons. This has no
impact on anything -unless- the user has specified a rankfile mapping
or is using the sequential mapper - in which case, the procs go to
the wrong place.

This patch correctly identifies the index of the node in the nodelist.