Per discussion with the v1.3 RMs, fudge the .so version numbers for libopen-rte.so and libopen-pal.so so that we don't break forward binary compatibility with MPI apps compiled since OMPI v1.3.2.
The real solution will have to come in trunk/v1.5 -- it's too late to
change how the wrapper compilers link in the v1.3 series.  Meaning:
even if we remove -lopen-rte and -lopen-pal in the v1.3 series,
there's still apps compiled with 1.3.2 and 1.3.3 that will need to
work.  Hence, we're locked into a "current" (the 1st number in the
c:r:a version number triple) value of 0 for the entire 1.3 and 1.4
series.  So changes to the wrappers here are useless -- might as well
start properly with 1.5.0, which will cause a binary compatibility
break between the 1.3/1.4 series and the 1.5/1.6 series.

Refs #2091 (see
http://www.open-mpi.org/community/lists/devel/2009/11/7081.php)