Fixes #2091.
After much back-n-forth this afternoon, the RM's have decided that in
order to get v1.3.4 out in a timely manner, we are punting on the
unexpectedly complex issue of versioning libopen-rte and libopen-pal
in v1.3.4.  Refs #2092 for more details.  Hopefully we can think of a
proper solution for v1.4.

Also advance the version number to 1.3.4rc4.