Fixes #2337: Move libltdl preopen patch to v1.5
r22806
Add a patch to libltdl so that the preopen loader doesn't mask a real
error if the file-based dloader (e.g., dlopen) fails to load a DSO for
a complex reason (e.g., unresolved symbol).

This has been widely reported upstream to the libltdl maintainers - a
general solution is difficult.  This is very definitely an
OMPI-specific solution.  Since our embedded libltdl is hidden behind
visibility flags, that's ok.

Note that this is a change to autogen.sh, but this commit does not
force re-running autogen.sh.  You'll just get the new functionality
the next time you re-run autogen.sh.

r22816
Oops -- use a name that matches the other patches in this dir and
matches the name in Makefile.am.

r22823
Fix for the libltdl preopen commit. So now autogen will progress without emiting an error.

{{{
grep: configure: No such file or directory
./autogen.sh: line 495: config/libltdl-preopen-error.patch: No such file or directory
}}}