Fixes #2677: Fix incorrect text in MPI_FILE_SET_VIEW man page
r24201
Correct an error in MPI_FILE_SET_VIEW man page: the fh does ''not''
have to have an identical ''value'' (it must be the same file handle,
but that means nothing about its actual value).  But the datarep must
have an identical value.  Additionaly, the etype does not need to be
identical, but the extent of all the etypes supplied must be
identical.

See MPI-2.2 p401-402 for further details.