Fixes #2719: Misleading MPI_BCAST error message
Thanks to Jeremiah Willcock for reporting this on the v1.4 branch. 

(patch submitted on #2719 based on r20907)

r20907
Correctly deal with the error case. The problem is tricky: the MPI
standard doesn't allow MPI_ERR_IN_STATUS to be returned from any
functions that return only one completed request (few exception here:
wait_some and wait_all and the test versions). As we use an wait_all
in these send_receive functions we should convert the
MPI_ERR_IN_STATUS to the real error, i.e. the one comming from the
MPI_ERROR field in the status corresponding to the failed request.