Fixes #2742: Move the new Winverbs BTL component into 1.5 branch
contrib/platform/win32/CMakeModules/opal_event_config.cmake file is missing
as it is not in the 1.5.

r24468
Add a variable for appending files into the list.

r24469
Install the config files.

r24477
Add the new WinVerbs btl component.

r24485
Fix the errors in Makefile, the source file list was not updated. Sorry about the MTT create failure.

r24487
Add svn:ignore and a configure.m4 that always fails (because this
component is only built via cmake... right?).

r24488
Missed another file for the tarball.