Fixes #2734: Move asm fix to v1.5
r24449
Fix register clobber list for x86 assembly.  Thanks to Jay Fenlason for the
patch.

r24453
Correct the assembly using xaddl for IA32.
Add atomic functions for add and sub 32 and 64 bits for AMD64.

r24528
Solve a problem when cross-compiling for PPC32 - in this case, OPAL_HAVE_ATOMIC_CMPSET_64 is not set, but the code requires that the ADD_64 and SUB_64 values at least be defined.