Fixes #2916: Pack error with partial memcopy
---svn-pre-commit-ignore-below---
Custom patch on ticket, based on r25488:

Fix an issue with the datatype engine where if the left over from
the previous pack was larger than the allowed space in the pack buffer
bad things happened. Thanks to Yuki MATSUMOTO and Takahiro Kawashima
for the bug report and the patch.