It's a little weird to put in a blank _SOURCES line for the HDFSFileFinder PROGRAM, but if we don't put in a _SOURCES line at all, Automake will default to "HDFSFileFinter_class_SOURCES = HDFSFileFinder.c", which clearly will cause problems.
But we don't want to put the .java file in _SOURCES, either, because
we haven't configured Automake to handle Java (because current
versions of Automake only have GCJ, not other Java compilers).  So set
HDFSFileFinder_class_SOURCES to blank and list the .java file in
EXTRA_SOURCES (so that they get picked up for "make dist").