Fixes #3105: Move ALLGATHERV fix to v1.6
---svn-pre-commit-ignore-below---
r26488
As reported by the mpi4py community (Bennet Fauber, Lisandro Dalcin,
Jonathan Dursi), fix a bug in the parameter verification of
MPI_ALLGATHERV: ensure to use the ''remote'' group size when checking
the recvcounts array.

r26489
Similar to r26488, also update the param error checking for
REDUCE_SCATTER, SCATTERV, and GATHERV.  Thanks to the mpi4py community
for reporting the issue.

r26490
Correctly use remote size where necessary. Completements the
patches for ticket #3105.

r26496
The sum of the local and remote recvcounts is identical (per MPI standard
requirement) so the recvcounts array is bounded by the number of
participants in the local group.