Minor fix for the command line parser: we didn't previously distinguish between unknown ''options'' (i.e., command line options that are registered and have some meaning) and unknown ''tokens'' (i.e., strings that do not begin with "-").  Hence, if you did:
   mpirun --fo my_mpi_program

(when perhaps you meant to type "--foo", mpirun would complain that no
such executable "--fo" existed.  That is ''correct,'' but perhaps not
completely useful.  It is more accurate for mpirun to report that
there is no such "--fo" option.

This change to cmd_line.c makes it so that we will ''always'' report
errors regarding tokens that begin with "-".