Fixes #3208: Update ctrl-c behavior in v1.7
---svn-pre-commit-ignore-below---
r26943
Ensure we can forcibly exit even when we are stuck inside of an event by replacing the libevent signal handler with a POSIX one that (a) attempts to trip a libevent termination event and (b) if anothe ctrl-c hits within 5 seconds, just calls exit.


r26945
A little cleaner on the message about repeated ctrl-c, and re-enable the event so we can abort if we see multiple ctrl-c's that don't meet the time requirement

r27005
Just a piece of this changeset - delete term_handler event upon normal termination.