Fixes #3241: Move minor coll tuned "err" assignment fix to v1.7
---svn-pre-commit-ignore-below---

r27070 [[BR]]
Ensure we assign "err" properly when invoking MCA_PML_CALLs.  Although
technically this is a necessary thing to do, it wasn't a tragedy that
we didn't have it because err was initialize to 0 in the beginning of
the functions where this problem occurred.  Also, OMPI will likely
abort if one of the MCA_PML_CALLs actually incurs an error (or, even
if it doesn't, MPI doesn't define the behavior anyway ;-) ).  

But looking forward to an FT-aware world, fixing this issue is a Good
Thing.  Many thanks to Hristo Iliev for pointing out the issue.